Move fetchFeatures
