+++ ee/client/lib/fetchFeatures.ts
